[Decomp][Spell] Cohesion split: extract trailing aura handlers to SpellAuraTrailing.cpp
Move the two handler groups that are physically stranded after the class-core
dividers (PeriodicTick/PeriodicDummyTick and the SpellAuraHolder block) into a
new SpellAuraTrailing.cpp translation unit:
 - Run 2: HandlePreventFleeing .. (the post-PeriodicDummyTick handler run), and
 - Tail: HandleModIncreaseSpellPowerPct, HandleOverrideSpellPowerByAp.

Pure file-level cohesion split: declarations stay in SpellAuras.h, the
AuraHandler[] dispatch table stays in SpellAuras.cpp and resolves the moved
member methods by address at link time. Two disjoint byte-identical deletions,
no behaviour change. SpellAuras.cpp 12994 -> 12608.
